Changing Interface for a base class
A lot of the logic for depositing and withdrawing may be similar, but this will not always be the case. To avoid code duplication as much as possible, moving from an interface to a base class which contains the default implementation of the deposit and withdraw methods, as-well as the default finance summary which should never need to be altered regardless of the account. The CalculateFinance method is likely the only method that will be unique across all of the different account types.