Starting to add Lifetime ISA Account
Firstly implementing a method to get the current tax year. Tax years work different to regular years, and run from the 6th April to the 5th April the next year. These are important for Lifetime ISA's because there is a maximum 4k deposit limit per tax year, and the end of the tax year is when the 25% bonus is added to the accounts balance.