Adding rough LISA calculations
This will require more testing and playing around with. Currently the 25% bonus is added flat to the account statement balance, which is simple but not accurate, however it there would be some complexity in adjusting this so that it is based on the end of tax years alone.

A potential way to solve this, would be to just loop through every transaction, for every tax year, sum them up, apply the 25%, then add these 25%; values together at the very end of the calculations method - when creating the AccountSummary.