Now calculating ISA bonus based on tax year
As explained in previous commit message, the ideal solution for fixing the tax year problem was to calculate the Lifetime ISA government 25% bonus based on the tax years themselves.

Now the bonus is only applied at the end of a tax year, which more accurately models the real world. Previously, when calculating your potential returns, even if you deposited 4k at the start of the tax year, and then calculated returns for the next day, the bonus would still be applied leaving an incorrect balance of 5k. Now the system can reflect that the government bonus would not have been applied at this stage. In the case described, depositing 4k at the start of the tax year, and then calculating potential returns based on the end date of the tax year - or future dates - will correctly add the government 25% bonus to the accounts balance in previous years.

The system also correctly models that government bonuses are not compounded.