Renaming GetTaxYearForTransaction
Renaming GetTaxYearForTransaction to GetTaxYearForDate. The previous name is misleading as it does not have a Transaction parameter, it instead takes a given date and returns the TaxYear object for that date, which contains information on the start and end date of the tax year that the provided date falls under.