Fixing Interest
Previously, interest was being added each month, and only started being compounded at the start of each new year. This is wrong, because if you open an account on in November, and then the interest compounds in January, you are a very lucky person.

Now fixing this to better reflect reality, interest can only be compounded on the accounts birthday, although it still accumulates and is added for each month.