Correcting Lifetime ISA Bonus Schedule
As I learn more about how ISA's work, I'm realising some of my earlier assumptions where wrong about how payments work. Initially I thought payments where paid at the end of the tax year, but there is actually a bonus period from the 5th of one month to the 6th of the next, and bonuses for that period are paid at the 28th. Adjusting the logic to represent this.

This actually affects my personal savings strategy, as it now makes more sense to pay immediately into a lifetime ISA to gain interest on the government bonus.