Improving InstantAccess interest calculations
Now improving InstantAccessAccount interest calculations to account for the fact that banks can chose when interest is both added to an account, and when it is compounded. The previous code just assumed that banks paid and then compounded interest each year, but in the real world this can differ - and is actually most likely to be interest paid monthly and compounded yearly.

This will need additional test cases around it, as-well as he LifetimeIsaAccount code being improved.